# e_learning_project-openapi.yaml
openapi: "3.1.0"
info:
  title: "e_learning_project API"
  description: "e_learning_project API"
  version: "1.0.0"
servers:
  - url: "https://e_learning_project"
paths:
  /orders:
    post:
      summary: "POST orders"
      operationId: "createOrder"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Order"
    get:
      summary: "GET orders"
      operationId: "orders"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      summary: "GET orders/{id}"
      operationId: "orderById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Order"
...
  /payments:
    post:
      summary: "POST payments"
      operationId: "payment"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
    OrderItemDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
    OrderRequestDTO:
      type: "object"
      properties:
        items:
          type: "array"
          items:
            $ref: "#/components/schemas/OrderItemDTO"
        discountCode:
          type: "string"
    MonetaryAmount:
      type: "object"
      properties: { }
    OrderItem:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        course:
          type: "integer"
          format: "int64"
        price:
          $ref: "#/components/schemas/MonetaryAmount"
    Instant:
      type: "object"
      properties: { }
    Order:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        items:
          type: "array"
          items:
            $ref: "#/components/schemas/OrderItem"
        orderDate:
          $ref: "#/components/schemas/Instant"
        totalPrice:
          $ref: "#/components/schemas/MonetaryAmount"
        discountedPrice:
          $ref: "#/components/schemas/MonetaryAmount"
        discountCode:
          type: "string"
        status:
          type: "string"
          enum:
            - "PENDING"
            - "PAID"
            - "CANCELLED"
        createdBy:
          type: "string"
        createdDate:
          $ref: "#/components/schemas/Instant"
        lastModifiedBy:
          type: "string"
        lastModifiedDate:
          $ref: "#/components/schemas/Instant"
...
    PaymentRequest:
      type: "object"
      properties:
        orderId:
          type: "string"
          format: "uuid"
        amount:
          $ref: "#/components/schemas/MonetaryAmount"
        paymentMethod:
          type: "string"
          enum:
            - "PAYPAL"
            - "STRIPE"
        token:
          type: "string"
    Payment:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        orderId:
          type: "string"
          format: "uuid"
        amount:
          $ref: "#/components/schemas/MonetaryAmount"
        status:
          type: "string"
          enum:
            - "PENDING"
            - "PAID"
            - "FAILED"
        paymentDate:
          type: "string"
          format: "date-time"
        paymentMethod:
          type: "string"
          enum:
            - "PAYPAL"
            - "STRIPE"
        transactionId:
          type: "string"
...